requirements.txt
tqdm==4.67.3
langchain==1.2.13
javascript==1!1.2.6
setuptools==65.5.0
openai==2.30.0
chardet==7.4.0.post2
# langchain-chroma 1.1.x requires chromadb>=1.3.5,<2. If api/types.py is incomplete, main.py runs bench/chromadb_bootstrap.py.
chromadb>=1.3.5,<2
tiktoken==0.12.0
requests==2.33.1
gymnasium==1.2.3
psutil==7.2.2
minecraft_launcher_lib==8.0
hydra-core==1.3.2
matplotlib==3.10.8
scipy==1.15.3
pandas==2.3.3
langchain-openai==1.1.12
langchain-chroma==1.1.0
langchain-ollama==1.0.1
